Remove raster class.
The original idea behind the raster class was to allow the state and
accumulator classes to use operating-system-specific routines to
optimize the layout of the 2D array; specifically, the vImage
framework of macOS.  Images could be passed into the PPHT library without the need for copying.

The problem was that the images still needed to be scanned, and the
translation layer turned out to be more expensive than originally
expected.  The raster class was eliminated in favor of including the
rasters directly into the aforementioned classes.